refinements, renamed components away from prorotype, validation item in in main sutdy secion nav bar now active properly
