+++ src/app/components/study/study.component.html
@@ -244,9 +244,9 @@
                 </li>
                 <li 
                 *ngIf="isCurator"
-                (click)="selectCurrentTab(7, 'prototype')"
+                (click)="selectCurrentTab(7, 'validationsv2')"
                 [ngClass]="{ 'is-active': (currentTabIndex$ | async) === 7}">
-                  Validation Prototype
+                  <a>Validation Framework V2</a>
 
                 </li>
               </ul>
@@ -306,7 +306,7 @@
                 <div
                   class="tab-content fadeIn animated"
                   *ngIf="(currentTabIndex$ | async) === 7">
-                  <validations-prototype></validations-prototype>
+                  <validations-v2-parent></validations-v2-parent>
 
 
                 </div>
