+++ src/app/components/study/study.module.ts
@@ -55,18 +55,18 @@ import { EuropePMCService } from "src/app/services/publications/europePMC.servic
 import { MatCheckboxModule } from "@angular/material/checkbox";
 import { RsyncStatusTransformPipe } from "./files/rsync/rsync-status-transform.pipe";
 import { RsyncComponent } from './files/rsync/rsync.component';
-import { ValidationStatusTransformPipe } from "./validations/validations-prototype/pipes/validation-status-transform.pipe";
-import { ValidationsPrototypeComponent } from "./validations/validations-prototype/validations-prototype/validations-prototype.component";
-import { ValidationInfoModalComponent } from "./validations/validations-prototype/validation-info-modal/validation-info-modal.component";
-import { RawViolationModalComponent } from "./validations/validations-prototype/raw-violation-modal/raw-violation-modal.component";
-import { PrototypeDetailComponent } from "./validations/validations-prototype/prototype-detail/prototype-detail.component";
+import { ValidationStatusTransformPipe } from "./validations/validations-v2/pipes/validation-status-transform.pipe";
 import { BrowserModule } from "@angular/platform-browser";
-import { ValidationReportSummaryComponent } from "./validations/validations-prototype/validation-report-summary/validation-report-summary.component";
-import { AddSpaceBeforeCapitalPipe } from "./validations/validations-prototype/pipes/add-space-before-capital.pipe";
-import { ValidationSectionComponent } from "./validations/validations-prototype/validation-section/validation-section.component";
-import { HandleUnderscoreInReportPipe } from "./validations/validations-prototype/pipes/handle-underscore-in-report.pipe";
-import { NoViolationsComponent } from "./validations/validations-prototype/misc/no-violations/no-violations.component";
-import { RemoveBackslashesPipe } from "./validations/validations-prototype/pipes/remove-backslashes.pipe";
+import { ValidationReportSummaryComponent } from "./validations/validations-v2/validation-report-summary/validation-report-summary.component";
+import { AddSpaceBeforeCapitalPipe } from "./validations/validations-v2/pipes/add-space-before-capital.pipe";
+import { ValidationSectionComponent } from "./validations/validations-v2/validation-section/validation-section.component";
+import { HandleUnderscoreInReportPipe } from "./validations/validations-v2/pipes/handle-underscore-in-report.pipe";
+import { NoViolationsComponent } from "./validations/validations-v2/misc/no-violations/no-violations.component";
+import { RemoveBackslashesPipe } from "./validations/validations-v2/pipes/remove-backslashes.pipe";
+import { ValidationInfoModalComponent } from "./validations/validations-v2/modals/validation-info-modal/validation-info-modal.component";
+import { RawViolationModalComponent } from "./validations/validations-v2/modals/raw-violation-modal/raw-violation-modal.component";
+import { ValidationV2DetailComponent } from "./validations/validations-v2/validation-detail/validation-detail.component";
+import { ValidationsV2ParentComponent } from "./validations/validations-v2/validations-v2-parent/validations-v2-parent.component";
 
 @NgModule({
   declarations: [
@@ -93,10 +93,10 @@ import { RemoveBackslashesPipe } from "./validations/validations-prototype/pipes
     RsyncStatusTransformPipe,
     ValidationStatusTransformPipe,
     RsyncComponent,
-    ValidationsPrototypeComponent,
+    ValidationsV2ParentComponent,
     ValidationInfoModalComponent,
     RawViolationModalComponent,
-    PrototypeDetailComponent,
+    ValidationV2DetailComponent,
     ValidationSectionComponent,
     NoViolationsComponent
   ],
@@ -159,10 +159,10 @@ import { RemoveBackslashesPipe } from "./validations/validations-prototype/pipes
     RsyncStatusTransformPipe,
     ValidationStatusTransformPipe,
     ValidationDetailCommentComponent,
-    ValidationsPrototypeComponent,
+    ValidationsV2ParentComponent,
     ValidationInfoModalComponent,
     RawViolationModalComponent,
-    PrototypeDetailComponent,
+    ValidationV2DetailComponent,
     ValidationSectionComponent,
     NoViolationsComponent
