+++ src/app/components/study/validations/validations-v2/interfaces/validation-report.types.ts
