+++ src/app/components/study/validations/validations-v2/misc/no-violations/no-violations.component.html
