+++ src/app/components/study/validations/validations-v2/modals/raw-violation-modal/raw-violation-modal.component.css
