+++ src/app/components/study/validations/validations-v2/modals/validation-info-modal/validation-info-modal.component.spec.ts
