+++ src/app/components/study/validations/validations-v2/pipes/add-space-before-capital.pipe.ts
