+++ src/app/components/study/validations/validations-v2/pipes/handle-underscore-in-report.pipe.ts
