+++ src/app/components/study/validations/validations-v2/pipes/remove-backslashes.pipe.ts
