+++ src/app/components/study/validations/validations-v2/pipes/validation-status-transform.pipe.ts
