+++ src/app/components/study/validations/validations-v2/validation-detail/validation-detail.component.css
@@ -132,6 +132,10 @@
     border-right: 2px solid #2f2f2f
 }
 
+.violation-footer-end-item {
+    padding: 5px;
+}
+
 .icon-container {
     display: inline-block; /* Ensure the container takes only as much space as needed */
     padding: 5px; /* Optional padding for better visuals */
