+++ src/app/components/study/validations/validations-v2/validation-detail/validation-detail.component.spec.ts
@@ -1,6 +1,6 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { PrototypeDetailComponent } from './prototype-detail.component';
+import { PrototypeDetailComponent } from './validation-detail.component';
 
 describe('PrototypeDetailComponent', () => {
   let component: PrototypeDetailComponent;
