+++ src/app/components/study/validations/validations-v2/validation-report-summary/validation-report-summary.component.spec.ts
