+++ src/app/components/study/validations/validations-v2/validation-section/validation-section.component.spec.ts
