+++ src/app/components/study/validations/validations-v2/validation-section/validation-section.component.ts
@@ -1,14 +1,6 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { ValidationReportSubsection } from '../interfaces/validation-report.types';
 import { Violation } from '../interfaces/validation-report.interface';
-import { MatButtonToggleGroup, MatButtonToggleModule } from '@angular/material/button-toggle';
-import { ValidationsPrototypeComponent } from '../validations-prototype/validations-prototype.component';
-import { StudyModule } from '../../../study.module';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { AddSpaceBeforeCapitalPipe } from '../pipes/add-space-before-capital.pipe';
-import { CommonModule } from '@angular/common';
-import { BrowserModule } from '@angular/platform-browser';
-import { MatButtonModule } from '@angular/material/button';
 
 @Component({
   selector: 'app-validation-section',
