+++ src/app/components/study/validations/validations-v2/validations-v2-parent/validations-v2-parent.component.css
