+++ src/app/components/study/validations/validations-v2/validations-v2-parent/validations-v2-parent.component.html
@@ -7,7 +7,7 @@
             <div *ngIf="allViolations.length > 0; else noViolations">
                 <div *ngIf="allViolations !== null">
                     <div *ngFor="let violation of allViolations; let i = index;">
-                        <prototype-detail [violation]="violation"></prototype-detail>
+                        <validation-v2-detail [violation]="violation"></validation-v2-detail>
                         <br>
                     </div>
                 </div>
